Fix problems with convert_biom_to_table
